edit datepicker and time slot for bookings

+++ app/Http/Controllers/UserController.php
@@ -5,39 +5,36 @@ namespace App\Http\Controllers;
 use App\Models\Role;
 use App\Models\User;
 use App\Models\Token;
-use App\Models\Permisions;
 use Illuminate\Http\Request;
-use App\Services\Notification;
-use Illuminate\Contracts\Session\Session;
-use Symfony\Component\HttpFoundation\Session\Session as SessionSession;
 
 class UserController extends Controller
 {
+    // Read
     public function index()
     {
         $users = User::all();
-
         $roles = Role::all();
-
         return view('admin.users.index', compact('users', 'roles'));
     }
 
-    public function profile($name)
+    public function profile($id)
     {
-        $user = User::where('name', $name)->first();
+        $user = User::where('id', $id)->first();
         return view('admin.users.profile', compact('user'));
     }
 
-//Create
-    public function create(){
+    // Create
+    public function create()
+    {
         $roles = Role::all();
         return view('admin.users.create', compact('roles'));
     }
-    public function store(Request $request){
-        $token = Token::where("user_phone", $request->phone)->first();
-        $token = $token->used;
 
-        if(!$token){
+    public function store(Request $request)
+    {
+        $token = Token::where("user_phone", $request->phone)->first();
+
+        if (!$token || $token->used) {
             return redirect()->back()->with('error', 'کد احراز هویت تایید نشده است.')->withInput();
         }
 
@@ -50,42 +47,43 @@ class UserController extends Controller
         ]);
 
         $user = User::create($request->only('name', 'email', 'password', 'phone'));
-
         $user->assignRole($request->role);
 
         return redirect()->route('users.index')->with("success", "کاربر جدید با موفقیت ایجاد شد.");
     }
 
-//Update
-    public function edit(User $user){
+    // Update
+    public function edit(User $user)
+    {
         $roles = Role::all();
         $user->load('roles');
-
         return view('admin.users.edit', compact('user', 'roles'));
     }
 
-    public function update(Request $request, User $user){
+    public function update(Request $request, User $user)
+    {
         $request->validate([
             'name' => 'required',
-            'email' => 'required|email|unique:users,email,'.$user->id,
+            'email' => 'required|email|unique:users',
+            'phone' => 'required|unique:users',
             'role' => 'required'
         ]);
 
         $user->refreshRoles($request->role);
-
         $user->update($request->only('name', 'email'));
 
         return redirect()->route('users.index')->with("success", "کاربر با موفقیت ویرایش شد.");
     }
-    //Update Phone
-    public function updatePhone(Request $request, User $user){
+
+    public function updatePhone(Request $request, User $user)
+    {
         $request->validate([
-            'phone' => 'required|unique:users,phone,'.$user->id,
+            'phone' => 'required|unique:users,phone,' . $user->id,
         ]);
 
         $token = Token::where("user_phone", $request->phone)->first();
 
-        if($token->used){
+        if ($token && $token->used) {
             $user->update($request->only('phone'));
             $token->delete();
 
@@ -95,16 +93,10 @@ class UserController extends Controller
         return redirect()->back()->with('error', 'کد احراز هویت تایید نشده است.')->withInput();
     }
 
-//Delete
-    public function destroy(User $user){
+    // Delete
+    public function destroy(User $user)
+    {
         $user->delete();
         return redirect()->route('users.index');
     }
-
-//Asign Role
-    // public function addRole(Request $request, User $user){
-    //     $user->assignRole($request->role);
-
-    //     return back()->with('success', true);
-    // }
 }
